requirements.txt
python-telegram-bot
aiohttp
aiohttp-socks
python-dotenv
ujson
